chore: rename _models.py to models.py
The module is 100% public API (RunResult, AgentInstallError,
AgentTimeoutError, TrajectorySource — all re-exported from benchflow).
The leading underscore was misleading: an agent landing in _models.py
sees a private file and assumes don't-depend, when in fact every symbol
is part of the documented public surface.

Updates all internal importers, the test_reexport module assertions,
the sdk.py architecture map, and the CLAUDE.md key-files / architecture
sections. _env_setup.py is intentionally left alone — it's a genuine
mix of public (stage_dockerfile_deps) and internal (_patch_harbor_dind)
symbols, so renaming the file would mislabel its internals.
