+++ tests/test_job.py
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from benchflow._models import RunResult
+from benchflow.models import RunResult
 from benchflow.job import Job, JobConfig, JobResult, RetryConfig
